feat: Validate base and new image dimensions
Before we even shell out, we need to validate if the images
we are comparing are of the same dimension, it used to be that
compare would error or mismatch dimensions. With newer versions
this no longer appears to be the case.

So we need to introduce this image container to validate
that both images are of equal dimensions before comparing.

This is expecially important for percentage threshold config
because the base image may have more pixels that the new image
causing it to incorrectly pass